[refac] 리소스 매핑 관련 함수를 api/exec/util로 이동
